debug

+++ binsearch.go
@@ -25,7 +25,7 @@ func (a sorter_uint64) Less(i, j int) bool { return a[i].k < a[j].k }
 func (f *Key_uint64) Find(thekey uint64) (uint64, bool) {
 	var min,at uint64
 	var current uint64
-	max := len(f.Key)-1
+	max := uint64(len(f.Key)-1)
 	for min<=max {
 		at = min+((max-min)/2)
 		if current=f.Key[at]; thekey<current {
@@ -47,7 +47,7 @@ func (f *Key_uint64) Find(thekey uint64) (uint64, bool) {
 // Find returns the index based on the key, using Interpolation.
 func (f *Key_uint64) FindInterpolation(thekey uint64) (uint64, bool) {
 	var l uint64
-	r := len(f.Key)-1
+	r := uint64(len(f.Key)-1)
 	var mid uint64 = uint64(((float32(thekey - f.Key[l])/float32(f.Key[r] - f.Key[l]))*float32(r))+0.5) // +0.5 makes it round instead of floor
 	for l<=r && mid>=l && mid<=r {
 		if (thekey < f.Key[mid]) {
@@ -118,7 +118,7 @@ func (a sorter_uint32) Less(i, j int) bool { return a[i].k < a[j].k }
 func (f *Key_uint32) Find(thekey uint32) (uint64, bool) {
 	var min,at uint64
 	var current uint32
-	max := len(f.Key)-1
+	max := uint64(len(f.Key)-1)
 	for min<=max {
 		at = (max+min)/2
 		if current=f.Key[at]; thekey<current {
@@ -140,7 +140,7 @@ func (f *Key_uint32) Find(thekey uint32) (uint64, bool) {
 // Find returns the index based on the key, using Interpolation.
 func (f *Key_uint32) FindInterpolation(thekey uint32) (uint64, bool) {
 	var l uint64
-	r := len(f.Key)-1
+	r := uint64(len(f.Key)-1)
 	var mid uint64 = uint64(((float32(thekey - f.Key[l])/float32(f.Key[r] - f.Key[l]))*float32(r))+0.5) // +0.5 makes it round instead of floor
 	for l<=r && mid>=l && mid<=r {
 		if (thekey < f.Key[mid]) {
@@ -210,7 +210,7 @@ func (a sorter_uint16) Less(i, j int) bool { return a[i].k < a[j].k }
 func (f *Key_uint16) Find(thekey uint16) (uint64, bool) {
 	var min,at uint64
 	var current uint16
-	max := len(f.Key)-1
+	max := uint64(len(f.Key)-1)
 	for min<=max {
 		at = (max+min)/2
 		if current=f.Key[at]; thekey<current {
@@ -232,7 +232,7 @@ func (f *Key_uint16) Find(thekey uint16) (uint64, bool) {
 // Find returns the index based on the key, using Interpolation.
 func (f *Key_uint16) FindInterpolation(thekey uint16) (uint64, bool) {
 	var l uint64
-	r := len(f.Key)-1
+	r := uint64(len(f.Key)-1)
 	var mid uint64 = uint64(((float32(thekey - f.Key[l])/float32(f.Key[r] - f.Key[l]))*float32(r))+0.5) // +0.5 makes it round instead of floor
 	for l<=r && mid>=l && mid<=r {
 		if (thekey < f.Key[mid]) {
@@ -302,7 +302,7 @@ func (a sorter_uint8) Less(i, j int) bool { return a[i].k < a[j].k }
 func (f *Key_uint8) Find(thekey uint8) (uint64, bool) {
 	var min,at uint64
 	var current uint8
-	max := len(f.Key)-1
+	max := uint64(len(f.Key)-1)
 	for min<=max {
 		at = (max+min)/2
 		if current=f.Key[at]; thekey<current {
@@ -324,7 +324,7 @@ func (f *Key_uint8) Find(thekey uint8) (uint64, bool) {
 // Find returns the index based on the key, using Interpolation.
 func (f *Key_uint8) FindInterpolation(thekey uint8) (uint64, bool) {
 	var l uint64
-	r := len(f.Key)-1
+	r := uint64(len(f.Key)-1)
 	var mid uint64 = uint64(((float32(thekey - f.Key[l])/float32(f.Key[r] - f.Key[l]))*float32(r))+0.5) // +0.5 makes it round instead of floor
 	for l<=r && mid>=l && mid<=r {
 		if (thekey < f.Key[mid]) {
@@ -393,7 +393,7 @@ func (a sorter_string) Less(i, j int) bool { return a[i].k < a[j].k }
 // Find returns the index based on the key.
 func (f *Key_string) Find(thekey string) (uint64, bool) {
 	var min,at uint64
-	max := len(f.Key)-1
+	max := uint64(len(f.Key)-1)
 	for min<=max {
 		at = min+((max-min)/2)
 		if thekey<f.Key[at] {
@@ -466,7 +466,7 @@ func (a sorter_bytes) Less(i, j int) bool { return bytes.Compare(a[i].k,a[j].k)=
 // Find returns the index based on the key.
 func (f *Key_bytes) Find(thekey []byte) (uint64, bool) {
 	var min,at uint64
-	max := len(f.Key)-1
+	max := uint64(len(f.Key)-1)
 	for min<=max {
 		at = min+((max-min)/2)
 		what := bytes.Compare(thekey,f.Key[at])
